requirements.txt
pandas==2.2.3
numpy==2.1.3
requests==2.32.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pyyaml==6.0.2
pytest==8.3.4
matplotlib==3.9.2
ipython==8.29.0